Removed AbstractPowBlockstore, functionality moved to AionBlockStore
