+++ modAionImpl/src/org/aion/zero/impl/db/AionBlockStore.java
@@ -29,7 +29,7 @@ import org.aion.log.AionLoggerFactory;
 import org.aion.log.LogEnum;
 import org.aion.mcf.blockchain.Block;
 import org.aion.mcf.blockchain.BlockHeader;
-import org.aion.mcf.db.AbstractPowBlockstore;
+import org.aion.mcf.db.IBlockStorePow;
 import org.aion.mcf.ds.DataSource;
 import org.aion.mcf.ds.DataSource.Type;
 import org.aion.mcf.ds.DataSourceArray;
@@ -43,7 +43,7 @@ import org.aion.util.conversions.Hex;
 import org.aion.zero.impl.types.AionBlock;
 import org.slf4j.Logger;
 
-public class AionBlockStore extends AbstractPowBlockstore {
+public class AionBlockStore implements IBlockStorePow {
 
     private static final Logger LOG = AionLoggerFactory.getLogger(LogEnum.DB.name());
     private static final Logger LOG_CONS = AionLoggerFactory.getLogger(LogEnum.CONS.name());
@@ -133,6 +133,22 @@ public class AionBlockStore extends AbstractPowBlockstore {
         } finally {
             lock.unlock();
         }
+    }
+
+    @Override
+    public byte[] getBlockHashByNumber(long blockNumber, byte[] branchBlockHash) {
+        Block branchBlock = getBlockByHash(branchBlockHash);
+        if (branchBlock.getNumber() < blockNumber) {
+            throw new IllegalArgumentException(
+                "Requested block number > branch hash number: "
+                    + blockNumber
+                    + " < "
+                    + branchBlock.getNumber());
+        }
+        while (branchBlock.getNumber() > blockNumber) {
+            branchBlock = getBlockByHash(branchBlock.getParentHash());
+        }
+        return branchBlock.getHash();
     }
 
     @Override
